removed server credentials
The server credentials were available to anyone, luckily no one hijacked the server or the account use to setup the server.
Also, the credentials have been changed, so anyone who does attempt to login with the previous revision of this README will be disappointed. If you want access to the account please email me at me@wgma00.me.